Fix bug in setting tip states and add test.
@dunleavy005 found that the tip CLV indexes weren't being used when
tip states were set. The initialization of the partition was relying
on the indexes of the tips returned by `pll_utree_query_tipnodes()`
being the same as their CLV indexes. I think this is the case for a
tree obtained directly from `pll_utree_parse_newick()`, but that can't
be expected in general.